fix: adapted the whole repository architecture to work with sync queue

+++ src/main/python/infrastructureServices/repositories/accessData/RemoteRepository.py
@@ -1,38 +1,46 @@
-import json
-from ...RestModule import executeCalls
-from ...RestModule import getCall
+from python.infrastructureServices.modules.RestModule import get_call
+
+
 class RemoteRepositoryImpl:
 
-    def __init__(self, serviceUrl, queue):
-        self.serviceUrl = serviceUrl
-        self.newRequest = {}
-        self.queue = queue
+    #TODO serve email e hash
+    def __init__(self, service_url, request_queue, hash):
+        self.hash = hash
+        self.service_url = service_url
+        self.request_queue = request_queue
 
-    def addElement(self, value):
-        self.newRequest = {
+    def add_element(self, user, value):
+        if user is "":
+            return
+        print("AGGIUNGO ELEMENTO DA INVIARE")
+        new_request = {
             "type": "PATCH",
-            "url": self.serviceUrl + "/userName/zaino1/cavolo1.json",
+            "url": self.service_url + "/" + user + "/" + self.hash + ".json",
             "payload": {
                 str(value): "true"
             }
         }
-        self.queue.put(self.newRequest)
-        req = self.queue.get()
-        element = json.loads(json.dumps(req))
-        executeCalls(element["type"], element["url"], element["payload"])
+        self.request_queue.put(new_request)
+        # req = self.request_queue.get()
+        # element = json.loads(json.dumps(req))
+        # execute_calls(element["type"], element["url"], element["payload"])
 
-    def removeElement(self, value):
-        self.newRequest = {
+    def remove_element(self, user, value):
+        if user is "":
+            return
+        new_request = {
             "type": "DELETE",
-            "url": self.serviceUrl + "/userName/zaino1/cavolo1/" + str(value) + ".json",
+            "url": self.service_url + "/" + user + "/" + self.hash + "/" + str(value) + ".json",
             "payload": {}
         }
-        self.queue.put(self.newRequest)
-        req = self.queue.get()
-        element = json.loads(json.dumps(req))
-        executeCalls(element["type"], element["url"], element["payload"])
+        self.request_queue.put(new_request)
+        # req = self.request_queue.get()
+        # element = json.loads(json.dumps(req))
+        # execute_calls(element["type"], element["url"], element["payload"])
         return
 
-    def findElement(self, value):
-        response = getCall(self.serviceUrl + "/userName/zaino1/cavolo1/" + str(value) + ".json")
+    def find_element(self, user, value):
+        if user is "":
+            return
+        response = get_call(self.service_url + "/" + user + "/" + self.hash + "/" + str(value) + ".json")
         return response.json() is not None
